my-companion/v1/requirements.txt
streamlit
langchain
openai
requests
sentence-transformers
faiss-cpu
python-docx
PyPDF2
tqdm
python-dotenv